requirements.txt
python-telegram-bot[job-queue]==21.*
openai>=1.30
httpx>=0.27
feedparser>=6.0
beautifulsoup4>=4.12
python-dotenv>=1.0
lxml>=5.0
